% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_bucket.R
\name{delete_bucket}
\alias{delete_bucket}
\title{Delete Bucket}
\usage{
delete_bucket(bucket, use_https, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{use_https}{True if connection is HTTPS and False if connection is HTTP}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if successful, \code{FALSE} otherwise.
}
\description{
Deletes an S3 bucket.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketDELETE.html}{API Documentation}
}
