% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Get Metadata}
\usage{
get_meta(path, bucket)
}
\arguments{
\item{path}{Path of the object in the s3 storage.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}
}
\value{
A dictionary of metadata pairs
}
\description{
Get metadata from an s3 object
}
\details{
Get tags from an s3 object
}
