% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get Tags}
\usage{
get_tags(path, bucket)
}
\arguments{
\item{path}{Path of the object in the s3 storage.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}
}
\value{
A dictionary of tag pairs
}
\description{
Get tags of an s3 object
}
\details{
Get tags from an s3 object
}
