% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_obj.R
\name{put_obj}
\alias{put_obj}
\title{Put Object}
\usage{
put_obj(object, bucket, file, meta)
}
\arguments{
\item{object}{A character string containing the name the object should have in S3 (i.e., its "object key"). If missing, the filename is used.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{file}{A character string containing the filename (or full path) of the file you want to upload to S3. Alternatively, an raw vector containing the file can be passed directly, in which case \code{object} needs to be specified explicitly.}

\item{meta}{A dictionary of metadata pairs}
}
\value{
If successful, \code{TRUE}
}
\description{
Puts an Object into an S3 storage
}
\details{
Puts an object into an S3 storage with metadata
}
