% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_tags.R
\name{put_tags}
\alias{put_tags}
\title{Put Tags}
\usage{
put_tags(path, bucket, tags)
}
\arguments{
\item{path}{Path of the object in the s3 storage.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{tags}{A dictionary of tag pairs}
}
\value{
If successful, \code{TRUE}
}
\description{
Puts tags in an s3 object
}
\details{
Puts tags into an s3 object
}
