% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_env.R
\name{setup_env}
\alias{setup_env}
\title{Setup environment}
\usage{
setup_env()
}
\description{
Setup environment variables in order to connect to an s3 storage
}
\details{
Setup of AWS_ACCESS_KEY, AWS_SECRET_ACCESS_KEY, AWS_DEFAULT_REGION nad AWS_S3_ENDPOINT
}
